+++ ImpactfulSkills/compatibility/Modcheck.cs
@@ -12,7 +12,13 @@ namespace ImpactfulSkills.compatibility {
         public static bool IsSNEAKerEnabled = false;
         public static bool IsMagicPluginEnabled = false;
         public static bool IsCrystalMagicalEnabled = false;
+
+        // Plant grid compatibility checks
         public static bool IsPlantEasilyEnabled = false;
+        public static bool IsGathilFarmGridEnabled = false;
+        public static bool IsVentureFarmGridEnabled = false;
+        public static bool IsXeioMassFarmingEnabled = false;
+        public static bool IsComfyMassFarmingEnabled = false;
 
         internal static void CheckModCompat() {
             try {
@@ -31,10 +37,29 @@ namespace ImpactfulSkills.compatibility {
                 if (plugins.Keys.Contains("advize.PlantEasily")) {
                     IsPlantEasilyEnabled = true;
                 }
+                if (plugins.Keys.Contains("fr.galathil.FarmGrid")) {
+                    IsGathilFarmGridEnabled = true;
+                }
+                if (plugins.Keys.Contains("com.orianaventure.mod.VentureFarmGrid")) {
+                    IsVentureFarmGridEnabled = true;
+                }
+                if (plugins.Keys.Contains("xeio.MassFarming")) {
+                    IsXeioMassFarmingEnabled = true;
+                }
+                if (plugins.Keys.Contains("EardwulfDoesMods.Comfy.MassFarming")) {
+                    IsComfyMassFarmingEnabled = true;
+                }
             } catch {
                 Logger.LogWarning("Unable to check mod compatibility. Ensure that Bepinex can load.");
             }
+        }
 
+        internal static bool OtherFarmingGridModPresent() {
+            if (IsPlantEasilyEnabled || IsGathilFarmGridEnabled || IsVentureFarmGridEnabled || IsXeioMassFarmingEnabled || IsComfyMassFarmingEnabled) {
+                return true;
+            }
+
+            return false;
         }
     }
 }